requirements.txt
requests
beautifulsoup4
python-dotenv
ollama
pdfplumber
python-docx
